refactor/#125 : LectureEvaluationTab 하드코딩 제거
