--- core/model/src/main/java/com/suwiki/core/model/enums/LectureEvaluationTab.kt
@@ -1,15 +1,0 @@
-package com.suwiki.core.model.enums
-
-enum class LectureEvaluationTab(
-  val position: Int,
-  val title: String,
-) {
-  LECTURE_EVALUATION(
-    position = 0,
-    title = "강의평가",
-  ),
-  EXAM_INFO(
-    position = 1,
-    title = "시험정보",
-  ),
-}
